+++ src/UI/SDLTextView.cpp
@@ -28,10 +28,10 @@ void SDLTextView::visualize() {
     long length = this->text.length();
 
     SDL_Rect Message_rect; //create a rect
-    Message_rect.x = (int)floorf((float)(posX * context->getTilewidth() * (context->getSCALE_FACTOR()))) + context->getX_offset();
-    Message_rect.y = (int) floorf((float)(posY * context->getTileHeigth() * (context->getSCALE_FACTOR())))+context->getY_offset();
-    Message_rect.w = (int)floor(0.6*length*fontSize*context->getSCALE_FACTOR());
-    Message_rect.h = (int)floor(0.6*fontSize*context->getSCALE_FACTOR());
+    Message_rect.x = (int)std::floor((float)(posX * context->getTilewidth() * (context->getSCALE_FACTOR()))) + context->getX_offset();
+    Message_rect.y = (int) std::floor((float)(posY * context->getTileHeigth() * (context->getSCALE_FACTOR())))+context->getY_offset();
+    Message_rect.w = (int)std::floor(0.6*length*fontSize*context->getSCALE_FACTOR());
+    Message_rect.h = (int)std::floor(0.6*fontSize*context->getSCALE_FACTOR());
     SDL_RenderCopy(context->getRenderer(), texture, nullptr, &Message_rect);
     SDL_DestroyTexture(texture);
     SDL_FreeSurface(surface);
